kafka: consolidate end txn handler
